net/dns: replace resolver IPs with type for DoH
We currently plumb full URLs for DNS resolvers from the control server
down to the client. But when we pass the values into the net/dns
package, we throw away any URL that isn't a bare IP. This commit
continues the plumbing, and gets the URL all the way to the built in
forwarder. (It stops before plumbing URLs into the OS configurations
that can handle them.)

Besides being a general cleanup for future features, this will let the
ipnlocal configure the DNS forwarder to use DNS-over-HTTP from the peer
API when a node is configured to use an exit node.

For #2596

Signed-off-by: David Crawshaw <crawshaw@tailscale.com>